Game 2048 : Add mix of variants

+++ src/game.hpp
@@ -27,7 +27,7 @@ protected:
 
 	vector<Sq>* board;
 
-	long score;
+	long long score;
 
 private:
 	bool quit;
